Etc/final-pjt/final-pjt-back/final_pjt_back/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
fixture==1.5.11
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
